dom/base/test/test_bug647518.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=647518
-->
<head>
  <title>Test for Bug 647518</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=647518">Mozilla Bug 647518</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 647518 **/
SimpleTest.waitForExplicitFinish();
var counter = 3;

var called = false;
var handle1 = window.requestAnimationFrame(function() {
  called = true;
});
ok(handle1 > 0, "Should get back a nonzero handle");

function checker() {
  --counter;
  if (counter == 0) {
    is(called, false, "Canceled callback should not have been called");
    SimpleTest.finish();
  } else {
    window.requestAnimationFrame(checker);
  }
}
window.requestAnimationFrame(checker);
window.cancelAnimationFrame(handle1);

</script>
</pre>
</body>
</html>
